Split .testing to 3 submodules; collect all fixtures and data
